Sentry fix: Don’t allow labels to be selected as the Gmail trash folder, see description
This is neat but Gmail has very specific semantics around labels vs the “spam”, “trash” and “all mail” folders. If you choose a label as your Trash the email still appears in all mail and Mailspring freaks out in several places. We could fix support for this scenarrio, but it’s unlikely this is what users actually want.