check all clipboard contents for Files, not just the first.
This was implemented for the "Copy Image" function of browsers, which on Firefox and Chrome on Linux was tested to give item[0] as "text/html" and item[1] as a file.
This effectively worked in rich text mode anyway as it is able to paste text/html snippets, but was unable to work in plain text mode.