Update fr.json
In the "Mail Rules" section, the two choices for "If %@ of the following conditions are met" are "All" and "Any", but the French translation is "Tout" (meaning "All") for both. I suggest replacing "Tout" (the current translation of "Any") by "N'importe lequel", but "Un seul" could also work. Warning: I don't know if "Any" is used for other buttons.