Downgrade to Electron 17.4.0 to fix Linux font rendering issue in Ubuntu 16 VM
The issue is fixed by passing —disable-gpu at launch, but it seems like it might be required for some users and the text is totally unlegible without the flag. Even the stock “electron” app exhibits the problem and it wasn’t required in Electron 17.4 so I’m hopeful they’ll fix it.