Fix date formatting issue and time not being set correctly (#2463)
The commit fixes a bug where the dateLikeString was not being formatted
correctly if it only contained 3 or 4 digits. The fix adds a check for
this case and formats the string accordingly.

Additionally, the commit addresses an issue where the time was not being
set correctly if the "gotTime" variable was false. The fix adds a check
for each iteration of the loop to ensure that the time is only set if
the corresponding "gotTime" value is false.