Fix slate issue caused by Chrome 106
Was able to determine that this was the problem by bisecting our Electron upgrade and identifying which chrome version bump broke it, and then googling for slate + chrome version.

It looks like this entire concept of webkit-user-modify is deprecated, so I have no idea why this matters, but it looks like they changed the default to read-write-plaintext and it broke Slate for everyone.

https://github.com/ianstormtaylor/slate/issues/5110