snap: Remove various unneeeded plugs
Remove various permissions that the snap required and that are not
actually used. In particular now all the file and settings access is handled
by XDG portals, so there's no need to grant direct file access.

Also this is not a media app or network management software, so no
need to add screen inhibitors or network observation.